#289 add swift3Action
