+++ core/swift3Action/proxy.py
@@ -29,7 +29,7 @@ proxy.debug = False
 
 SRC_EPILOGUE_FILE = "./epilogue.swift"
 DEST_SCRIPT_FILE = "/swift3Action/spm-build/main.swift"
-DEST_SCRIPT_DIR = "/swif3Action/spm-build"
+DEST_SCRIPT_DIR = "/swift3Action/spm-build"
 DEST_BIN_FILE = "/swift3Action/spm-build/.build/debug/Action"
 BUILD_PROCESS = [ "swift", "build", "-Xcc", "-fblocks", "-Xlinker", "-rpath", "-Xlinker", "/swift3Action/spm-build/.build/debug"]
 #BUILD_PROCESS = [ "swift", "build", "-Xcc", "-fblocks"]
